PR review : Add DrawSmallTexture to the MaterialEditorUtility class.
